crud completed

+++ src/main/java/com/crud/crudproject/repositories/LeadRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.crud.crudproject.entities.Lead;
 
-public interface LeadRepository extends JpaRepository<Lead ,Integer>{
+public interface LeadRepository extends JpaRepository<Lead ,Long>{
 
 }
